Changed schedule generator and added violation checks
Added checks to make sure the TTP schedules don't violate the three TTP constraints. Also changed the generator so the matchups won't have to be passed on all the time. This was done by adding a function which removes all possible future rounds containing the at least one of the same matchups as the current round.