Added/updated comments and folder restructure
Added some new comments to new code, updated some comments so they are worded a bit nicer. Removed old commented code.
Also added a new folder for all schedules. So now all folders containing schedules from the same run/with the same name are in a folder together (as before), and these folders are now in a folder called Schedules.